+++ common/src/main/java/com/leobeliik/extremesoundmuffler/SoundMufflerCommon.java
@@ -2,7 +2,6 @@ package com.leobeliik.extremesoundmuffler;
 
 import com.leobeliik.extremesoundmuffler.gui.MufflerScreen;
 import com.mojang.blaze3d.platform.InputConstants;
-import com.mojang.blaze3d.systems.RenderSystem;
 import net.minecraft.client.KeyMapping;
 import net.minecraft.network.chat.contents.TranslatableContents;
 import net.minecraft.resources.ResourceLocation;
@@ -10,10 +9,6 @@ import net.minecraft.resources.ResourceLocation;
 public class SoundMufflerCommon {
 
     //whether should use the dark or light textures
-    public static void renderGui() {
-        RenderSystem.setShaderTexture(0, getTextureRL());
-    }
-
     public static ResourceLocation getTextureRL() {
         String texture = CommonConfig.get().useDarkTheme().get() ? "textures/gui/sm_gui_dark.png" : "textures/gui/sm_gui.png";
         return ResourceLocation.fromNamespaceAndPath(Constants.MOD_ID, texture);