+++ common/src/main/java/com/leobeliik/extremesoundmuffler/gui/buttons/InvButton.java
@@ -9,6 +9,8 @@ import net.minecraft.client.gui.narration.NarratedElementType;
 import net.minecraft.client.gui.narration.NarrationElementOutput;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.network.chat.Component;
+import org.jetbrains.annotations.NotNull;
+
 import static com.leobeliik.extremesoundmuffler.SoundMufflerCommon.getTextureRL;
 
 public class InvButton extends AbstractButton implements IColorsGui {
@@ -26,9 +28,8 @@ public class InvButton extends AbstractButton implements IColorsGui {
     }
 
     @Override
-    public void renderScrollingString(GuiGraphics render, Font font, int mouseX, int mouseY) {
+    public void renderScrollingString(@NotNull GuiGraphics render, @NotNull Font font, int mouseX, int mouseY) {
         if (this.visible) {
-            SoundMufflerCommon.renderGui();
             render.blit(RenderType::guiTextured, getTextureRL(), getX(), getY(), 43f, 202f, 11, 11, 256, 256); //button texure
             if (isMouseOver(mouseX, mouseY) && !hold) {
                 render.drawCenteredString(font, Component.translatable("inventory.btn"), getX() + 5, getY() + this.height + 1, whiteText);