+++ common/src/main/java/com/leobeliik/extremesoundmuffler/mixins/SoundMixin.java
@@ -17,6 +17,8 @@ import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
+import java.util.Locale;
+
 @Mixin(SoundEngine.class)
 public abstract class SoundMixin implements ISoundLists {
 
@@ -61,6 +63,7 @@ public abstract class SoundMixin implements ISoundLists {
             }
             float tempVolume = tempSound.getVolume();
             String soundName = soundResourceLocation.getPath();
+            String modName = soundResourceLocation.getNamespace();
 
             //global sounds like thunder or dragon growl has too high volume to be properly muffled, so first we temporarily lower the max volume
             if (soundName.contains("entity.lightning_bolt.thunder") || soundName.contains("entity.ender_dragon.growl")) {
@@ -78,6 +81,14 @@ public abstract class SoundMixin implements ISoundLists {
                     return (float) (tempVolume * anchor.getMuffledSounds().get(soundResourceLocation));
                 }
             }
+
+            //Mod wide muffling from config
+            for (String mod : modsMuffled) {
+                if (mod.contains(modName.toLowerCase(Locale.ROOT))) {
+                    return tempVolume * Float.parseFloat(mod.split(":")[1]);
+                }
+            }
+
         }
 
         return volume;
@@ -97,5 +108,4 @@ public abstract class SoundMixin implements ISoundLists {
     private static boolean esm_isForbidden(SoundInstance sound) {
         return forbiddenSounds.stream().anyMatch(fs -> sound.getLocation().toString().contains(fs));
     }
-
 }