+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-Fix mod trying to load on server side
+Fixed "forbiddenSounds" config option preventing sounds from being muffled instead of just hide them from the "recent sounds list"